Reorganize code into mindmap and worklog packages
**mindmap package**: Core mindmap reading and tree traversal
- mindmap/reader.py: NodeTreeHelper, DateReader, DateTimeReader
- mindmap/models.py: DateValue, DateTimeValue, TimeEntry, Section, DateEntry

**worklog package**: Worklog-specific formatting and TODO logic
- worklog/format.py: TodoHelper (is_todo, clean_todo_text)
- worklog/helpers.py: DateTimeHelper, DurationFormatter, HierarchicalNodeProcessor
- worklog/models.py: TaskEntry, TaskInfo, ProjectInfo

**Updates to formatters**:
- orgmode.py: Import from mindmap.reader, worklog.helpers, worklog.format
- orgmode_date_sections.py: Import from mindmap.reader, mindmap.models, worklog.format
- orgmode_lists.py: Import from mindmap.reader, worklog.format

**Benefits**:
- Clear separation of concerns (mindmap vs worklog)
- Easier to extend with new formatters or readers
- Better code organization following single responsibility principle
- All 203 tests pass with strict mypy type checking
